Feature/colenda 21 metadata source objects

+++ app/helpers/metadata_builder_helper.rb
@@ -1,15 +1,7 @@
 module MetadataBuilderHelper
 
-  def render_form_or_message(partial)
-    if @object.metadata_builder.source.present?
-      render :partial => partial, :locals => {metadata_builder: @object.metadata_builder}
-    else
-      render :partial => "metadata_builders/no_source"
-    end
-  end
-
   def render_preview_xml
-    if @object.metadata_builder.preserve.present?
+    if MetadataSource.where(:metadata_builder_id => @object.metadata_builder.id).pluck(:user_defined_mappings).present?
       render :partial => "metadata_builders/preview_xml"
     else
       render :partial => "metadata_builders/no_mappings"
@@ -24,14 +16,6 @@ module MetadataBuilderHelper
     end
   end
 
-  def render_metadata_generation_form
-    if @object.metadata_builder.source.present?
-      render :partial => "metadata_builders/form"
-    else
-      render :partial => "metadata_builders/no_source"
-    end
-  end
-
   def render_source_select_form
     if @object.metadata_builder.available_metadata_files.present?
       render :partial => "metadata_builders/source_select"
@@ -40,27 +24,23 @@ module MetadataBuilderHelper
     end
   end
 
-  def render_source_specs_form
-    if @object.metadata_builder.source.present?
-      render :partial => "metadata_builders/source_specs"
-    end
-  end
-
   def render_sample_xml
     @object.version_control_agent.clone
     @object.version_control_agent.get(:get_location => "#{@object.version_control_agent.working_path}/#{@object.metadata_subdirectory}")
     @sample_xml_docs = ""
     @file_links = Array.new
-    Dir.glob("#{@object.version_control_agent.working_path}/#{@object.metadata_subdirectory}/*.xml").each do |file|
-      @file_links << link_to(_prettify(file), "##{file}")
-      anchor_tag = content_tag(:a, "", :name=> file)
-      sample_xml_content = File.open(file, "r"){|io| io.read}
-      sample_xml_doc = REXML::Document.new sample_xml_content
-      sample_xml = ""
-      sample_xml_doc.write(sample_xml, 1)
-      header = content_tag(:h3, "XML Sample for #{_prettify(file)}")
-      xml_code = content_tag(:pre, "#{sample_xml}")
-      @sample_xml_docs << content_tag(:div, anchor_tag << header << xml_code, :class => "doc")
+    @object.metadata_builder.preserve.each do |file|
+      if File.exist?(file)
+        @file_links << link_to(prettify(file), "##{file}")
+        anchor_tag = content_tag(:a, "", :name=> file)
+        sample_xml_content = File.open(file, "r"){|io| io.read}
+        sample_xml_doc = REXML::Document.new sample_xml_content
+        sample_xml = ""
+        sample_xml_doc.write(sample_xml, 1)
+        header = content_tag(:h3, "XML Sample for #{prettify(file)}")
+        xml_code = content_tag(:pre, "#{sample_xml}")
+        @sample_xml_docs << content_tag(:div, anchor_tag << header << xml_code, :class => "doc")
+      end
     end
     @object.version_control_agent.delete_clone
     @file_links_html = ""
@@ -70,28 +50,7 @@ module MetadataBuilderHelper
     return content_tag(:ul, @file_links_html.html_safe) << @sample_xml_docs.html_safe
   end
 
-  def _structural_elements(file_name)
-    root_default = ""
-    child_default = ""
-    if @object.metadata_builder.field_mappings.present?
-      root_element = @object.metadata_builder.field_mappings[file_name]["root_element"]["mapped_value"].present? ? @object.metadata_builder.field_mappings[file_name]["root_element"]["mapped_value"] : root_default
-      child_element = @object.metadata_builder.field_mappings[file_name]["child_element"]["mapped_value"].present? ? @object.metadata_builder.field_mappings[file_name]["child_element"]["mapped_value"] : child_default
-      return root_element, child_element
-    else
-      return root_default, child_default
-    end
-  end
-
-  def _nested_relationships_values(parent_file)
-    child_array = Array.new
-    child_candidates = _prettify(@object.metadata_builder.source)
-    child_candidates.each do |child|
-      child_array << [child, { parent_file => child }.to_s] unless _prettify(parent_file) == child
-    end
-    return child_array
-  end
-
-  def _prettify(file_path_input)
+  def prettify(file_path_input)
       if file_path_input.is_a? Array
         file_path_array = Array.new
         file_path_input.each do |file_path|
@@ -102,9 +61,11 @@ module MetadataBuilderHelper
         file_path_string = _prettified_working_file(file_path_input)
         return file_path_string
       else
-        raise "Invalid argument #{file_path_input}. _prettify can only accept strings and arrays of strings."
+        raise "Invalid argument #{file_path_input}. prettify can only accept strings and arrays of strings."
       end
   end
+
+  private
 
   def _prettified_working_file(file_path)
     return file_path.gsub(@object.version_control_agent.working_path, "")
